requirements.txt
flask
python-dotenv
groq
PyPDF2
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
duckduckgo-search

